end of mearging plotters

+++ IpCHashT.hpp
@@ -23,8 +23,8 @@ namespace sstd{
 			  class T_val,
 			  class T_hash   = std::hash<T_key>,
 			  class T_key_eq = std::equal_to<T_key>,
-//			  typename T_shift = uint8
-			  typename T_shift = uint16
+			  typename T_shift = uint8
+//			  typename T_shift = uint16
 			  >
 	class IpCHashT; // chained hash table
 }
